feat(macos): Clean restart — CMake build system + Platform layer + Win32 shims
Build system:
- CMakePresets.json: macOS and macOS-release presets (Ninja, Debug, ARM64)
- Root CMakeLists.txt: Apple branches for DX8/GameSpy/Miles/Bink/Win32 dummy libs
- Platform/MacOS/CMakeLists.txt: Metal backend, Cocoa integration, platform stubs
- GeneralsMD/Code/Main/CMakeLists.txt: macOS entry point (MacOSMain.mm)

Win32 compatibility shims (Platform/MacOS/Include/):
- windows.h: Base types (DWORD, HANDLE, BOOL, RECT, etc.), MSVC aliases
- objbase.h: COM stubs (IUnknown, GUID, HRESULT)
- malloc.h: Redirects to stdlib.h (macOS has no separate malloc.h)

CMake fixes (exclude, don't stub):
- debug_stack.cpp: Excluded on Apple (Win32 StackWalk64)
- DbgHelp*.cpp: Excluded on Apple (imagehlp.h)
- PCH ordering: windows.h moved before game headers in all PCH lists
- osdep.h: Restored for _UNIX builds (official project mechanism)
- Dependencies/Utility: Added Utility/ subdir to include path on non-Windows

Platform layer (57 files from feature/macos-support):
- Metal/: DX8 → Metal rendering backend (MetalDevice8, buffers, textures)
- Main/: MacOSMain, GameClient, WindowManager, Input, D3DXStubs
- Client/: Display, DisplayString (CoreText)
- Audio/: MacOSAudioManager (AVAudioEngine)
- Common/: Cross-platform .big + local file system
- Stubs/: GameSpy, LZHL, WWDownload, GitInfo, W3DShaderManager

Status: 13 files still failing (WW3D2 Win32 API calls, NoxCompress missing stdio)

+++ Dependencies/Utility/Utility/thread_compat.h
@@ -16,18 +16,12 @@
 **	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-// This file contains thread related functions for compatibility with non-windows platforms.
+// This file contains thread related functions for compatibility with
+// non-windows platforms.
 #pragma once
 #include <pthread.h>
 #include <unistd.h>
 
-inline int GetCurrentThreadId()
-{
-  return pthread_self();
-}
+inline int GetCurrentThreadId() { return (int)(uintptr_t)pthread_self(); }
 
-inline void Sleep(int ms)
-{
-  usleep(ms * 1000);
-}
-
+inline void Sleep(int ms) { usleep(ms * 1000); }
